requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
redis==5.2.0
celery==5.4.0
pydantic-settings==2.7.0
PyJWT==2.10.1
argon2-cffi==23.1.0
langgraph==0.2.60
pandas==2.2.3
scikit-learn==1.5.2

pytest==8.3.4
httpx==0.28.1
